core: Use uintptr_t instead for DISABLE_ANN_PAIR_OPTIMIZATION

+++ src/Pair.h
@@ -206,7 +206,7 @@ struct Pair
     Object car;
     Object cdr;
 #ifdef DISABLE_ANN_PAIR_OPTIMIZATION
-    uint64_t space;
+    uintptr_t space;
 #endif
 };
